+++ src/frontend/packages/lib_video/src/api/useTranscriptReaderRequest/index.ts
@@ -1,32 +1,22 @@
 import { fetchWrapper } from 'lib-components';
 import { useQuery, UseQueryOptions } from 'react-query';
-import { VTTCue, WebVTT } from 'vtt.js';
 
 export const useTranscriptReaderRequest = (
+  id: string,
   url: string,
-  onSuccess: (cue: VTTCue) => void,
   queryConfig?: UseQueryOptions<string, 'useTranscriptReaderRequest', string>,
 ) => {
   return useQuery({
-    queryKey: ['useTranscriptReaderRequest'],
+    queryKey: ['useTranscriptReaderRequest', id],
     queryFn: async () => {
       const response = await fetchWrapper(url);
       return await response.text();
     },
-    onSuccess: (transcriptReader: string) => {
-      if (!transcriptReader) {
-        return;
-      }
-
-      const parser = new WebVTT.Parser(window, WebVTT.StringDecoder());
-
-      parser.oncue = (cue: VTTCue) => {
-        onSuccess(cue);
-      };
-
-      parser.parse(transcriptReader);
-      parser.flush();
-    },
+    refetchInterval: false,
+    refetchIntervalInBackground: false,
+    refetchOnWindowFocus: false,
+    cacheTime: Infinity,
+    staleTime: Infinity,
     ...queryConfig,
   });
 };